requirements.txt
aiosqlite>=0.21.0
aiohttp>=3.13.2
aiofiles>=24.1.0
python-dotenv>=1.2.1
maxapi>=0.9.7
gigachat
